organising Nominatim, (duplicate code!) and introducing a mapViewModel for the markers
