+++ app/src/main/java/com/monkeyteam/chimpagne/model/location/Nominatim.kt
@@ -2,24 +2,36 @@ package com.monkeyteam.chimpagne.model.location
 
 import android.util.Log
 import java.io.IOException
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import okhttp3.Call
 import okhttp3.Callback
+import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
+import okhttp3.Request
 import okhttp3.Response
 import org.json.JSONArray
 
-class LocationHelper {
+object NominatimConstants {
+  const val SCHEME = "https"
+  const val HOST = "nominatim.openstreetmap.org"
+}
 
-  private val _markers = MutableStateFlow<List<Location>>(emptyList())
-  val markers: StateFlow<List<Location>> = _markers.asStateFlow()
+class Nominatim {
 
-  suspend fun convertNameToLocation(name: String, onResult: (Location) -> Unit) {
+  fun convertNameToLocation(name: String, onResult: (List<Location>) -> Unit, limit: Int = 5) {
+
     val client = OkHttpClient()
-    val urlSearchNominatim = "https://nominatim.openstreetmap.org/search?q=$name&format=json"
-    val request = okhttp3.Request.Builder().url(urlSearchNominatim).build()
+
+    val url =
+        HttpUrl.Builder()
+            .scheme(NominatimConstants.SCHEME)
+            .host(NominatimConstants.HOST)
+            .addPathSegment("search")
+            .addQueryParameter("q", java.net.URLEncoder.encode(name, "UTF-8"))
+            .addQueryParameter("format", "json")
+            .addQueryParameter("limit", limit.toString())
+            .build()
+
+    val request = Request.Builder().url(url).addHeader("User-Agent", "Chimpagne").build()
     client
         .newCall(request)
         .enqueue(
@@ -30,7 +42,7 @@ class LocationHelper {
                     "LocationHelper",
                     "Failed to get location for $name, because of IO Exception",
                     e)
-                onResult(Location(name))
+                onResult(listOf())
               }
 
               override fun onResponse(call: Call, response: Response) {
@@ -39,37 +51,28 @@ class LocationHelper {
                     Log.e(
                         "LocationHelper",
                         "Failed to get location for $name, because of unsuccessful response")
-                    onResult(Location(name))
+                    onResult(listOf())
                   } else {
                     val geoLocation = response.body?.string()
                     val jsonArray = geoLocation?.let { JSONArray(it) }
                     if (jsonArray != null && jsonArray.length() > 0) {
-                      val jsonObject = jsonArray.getJSONObject(0)
-                      val lat = jsonObject.getDouble("lat")
-                      val lon = jsonObject.getDouble("lon")
-                      val locationInfo = Location(name, lat, lon)
-                      onResult(locationInfo)
+                      val locations = arrayListOf<Location>()
+                      for (i in 0 ..< jsonArray.length()) {
+                        val jsonObject = jsonArray.getJSONObject(i)
+                        val lat = jsonObject.getDouble("lat")
+                        val lon = jsonObject.getDouble("lon")
+                        locations.add(Location(name, lat, lon))
+                      }
+                      onResult(locations)
                     } else {
                       Log.e(
                           "LocationHelper",
                           "Failed to get location for $name, because of empty response")
-                      onResult(Location(name))
+                      onResult(listOf())
                     }
                   }
                 }
               }
             })
-  }
-
-  suspend fun addMarker(location: Location) {
-    val currentMarkers = _markers.value.toMutableList()
-    currentMarkers.add(location)
-    _markers.value = currentMarkers
-  }
-
-  fun removeMarker(location: Location) {
-    val currentMarkers = _markers.value.toMutableList()
-    currentMarkers.removeAll { it.name == location.name }
-    _markers.value = currentMarkers
   }
 }
